requirement.txt
torch
opencv-python
torchvision
fastapi
loguru
httpx
uvicorn
pydantic-settings